requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-generativeai
yfinance
pydantic
